Check read success
As reported by https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=16960 and https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=16878

+++ src/matio_private.h
@@ -140,31 +140,31 @@ EXTERN mat_int16_t   Mat_int16Swap(mat_int16_t  *a);
 EXTERN mat_uint16_t  Mat_uint16Swap(mat_uint16_t *a);
 
 /* read_data.c */
-EXTERN int ReadDoubleData(mat_t *mat,double *data,enum matio_types data_type,
+EXTERN size_t ReadDoubleData(mat_t *mat,double *data,enum matio_types data_type,
                size_t len);
-EXTERN int ReadSingleData(mat_t *mat,float  *data,enum matio_types data_type,
+EXTERN size_t ReadSingleData(mat_t *mat,float  *data,enum matio_types data_type,
                size_t len);
 #ifdef HAVE_MAT_INT64_T
-EXTERN int ReadInt64Data (mat_t *mat,mat_int64_t *data,
+EXTERN size_t ReadInt64Data (mat_t *mat,mat_int64_t *data,
                enum matio_types data_type,size_t len);
 #endif /* HAVE_MAT_INT64_T */
 #ifdef HAVE_MAT_UINT64_T
-EXTERN int ReadUInt64Data(mat_t *mat,mat_uint64_t *data,
+EXTERN size_t ReadUInt64Data(mat_t *mat,mat_uint64_t *data,
                enum matio_types data_type,size_t len);
 #endif /* HAVE_MAT_UINT64_T */
-EXTERN int ReadInt32Data (mat_t *mat,mat_int32_t *data,
+EXTERN size_t ReadInt32Data (mat_t *mat,mat_int32_t *data,
                enum matio_types data_type,size_t len);
-EXTERN int ReadUInt32Data(mat_t *mat,mat_uint32_t *data,
+EXTERN size_t ReadUInt32Data(mat_t *mat,mat_uint32_t *data,
                enum matio_types data_type,size_t len);
-EXTERN int ReadInt16Data (mat_t *mat,mat_int16_t *data,
+EXTERN size_t ReadInt16Data (mat_t *mat,mat_int16_t *data,
                enum matio_types data_type,size_t len);
-EXTERN int ReadUInt16Data(mat_t *mat,mat_uint16_t *data,
+EXTERN size_t ReadUInt16Data(mat_t *mat,mat_uint16_t *data,
                enum matio_types data_type,size_t len);
-EXTERN int ReadInt8Data  (mat_t *mat,mat_int8_t  *data,
+EXTERN size_t ReadInt8Data  (mat_t *mat,mat_int8_t  *data,
                enum matio_types data_type,size_t len);
-EXTERN int ReadUInt8Data (mat_t *mat,mat_uint8_t  *data,
+EXTERN size_t ReadUInt8Data (mat_t *mat,mat_uint8_t  *data,
                enum matio_types data_type,size_t len);
-EXTERN int ReadCharData  (mat_t *mat,char  *data,enum matio_types data_type,
+EXTERN size_t ReadCharData  (mat_t *mat,char  *data,enum matio_types data_type,
                int len);
 EXTERN int ReadDataSlab1(mat_t *mat,void *data,enum matio_classes class_type,
                enum matio_types data_type,int start,int stride,int edge);
